fixed /parts/list

+++ src/main/java/com/mercadolibre/fernandez_federico/repositories/IStockRepository.java
@@ -4,5 +4,5 @@ import com.mercadolibre.fernandez_federico.models.Stock;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IStockRepository extends JpaRepository<Stock, Integer>
+public interface IStockRepository extends JpaRepository<Stock, Long>
 { }
